Remove hardcoded password and switch to env variable

+++ check_duplicates.py
@@ -5,6 +5,9 @@ from google.auth.transport.requests import Request
 from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import InstalledAppFlow
 from googleapiclient.discovery import build 
+import smtplib
+from email.message import EmailMessage
+import os
 
 SCOPES = ['https://www.googleapis.com/auth/drive']
 
@@ -30,12 +33,12 @@ def main():
     results = service.files().list(
         pageSize=1000, fields="nextPageToken, files(id, name, md5Checksum)").execute()
     items = results.get('files', [])
-    if not items:
-        print('No files found.')
-    else:
-        print('Files:')
-        for item in items:
-            print(f"{item['name']} ({item['id']})")
+    # if not items:
+    #     print('No files found.')
+    # else:
+    #     print('Files:')
+    #     for item in items:
+    #         print(f"{item['name']} ({item['id']})")
     
     seen = {}
     duplicates = []
@@ -57,18 +60,35 @@ def main():
             duplicates.append(file)
         else:
             seen[key] = file['id']
-    print("Found duplicates")
-    confirm = input("Delete duplicates? (yes/no): ")
-    if confirm.lower() == "yes":
-        for dup in duplicates:
-            file_id = dup['id']
-            try:
-                service.files().delete(fileId=file_id).execute()
-                print(f"Deleted: {dup['name']} ({file_id})")
-            except Exception as e:
-                print(f"Failed to delete {dup['name']}: {e}")
+    if duplicates:
+        confirm = input("Delete duplicates? (yes/no): ")
+        if confirm.lower() == "yes":
+            deleted_files = []
+            for dup in duplicates:
+                file_id = dup['id']
+                try:
+                    service.files().delete(fileId=file_id).execute()
+                    print(f"Deleted: {dup['name']} ({file_id})")
+                    deleted_files.append(dup['name'])  
+                except Exception as e:
+                    print(f"Failed to delete {dup['name']}: {e}")
+            
+            if deleted_files:
+                #create the email
+                msg = EmailMessage()
+                msg['Subject'] = 'Duplicate files deleted'
+                msg['From'] = 'Python_duplicates_remover'
+                msg['To'] = 'houriahasbell@gmail.com'
+                msg.set_content(f"The following files were deleted because they were duplicates:\n\n" + "\n".join(deleted_files))
+    
+                # Send via Gmail SMTP
+                with smtplib.SMTP_SSL('smtp.gmail.com', 465) as smtp:
+                    smtp.login('houriahasbell@gmail.com', os.getenv("EMAIL_APP_PASSWORD"))
+                    smtp.send_message(msg)
+        else:
+            print("No files deleted.")
     else:
-        print("No files deleted.")
+        print('no duplicate founds')
 if __name__ == '__main__':
     main()
